introduce
hello friends my name is tri hadi suwanto i am a new writer who needs to come from a remote city in lampung. I am often called Hadi. I am 28 years old and I have a wife and one child who is still 3 years old.I like singing the most, even though my voice is not very melodious, but I always study with enthusiasm. besides that I love fishing usually I fish in the lake in the area we live. I always keep my simple family and harmony in my family is the main thing. I like to spend time off on tourist tours in my area. I like the flower garden that makes us and our family feel comfortable in our lives, because the comfort and tranquility of life is what we are looking for.

my Hobby

this flower park is located in a metro city not far from where we live. friends are obliged to stop by when traveling to Lampung.because friends, if you have visited this park, are life burdens or problems that you feel will slowly find a solution because whatever the problem is, if we are calm, think the solution will come out in our minds. tourism in Lampung is currently experiencing very rapid progress, starting from entertainment, playgrounds, beaches, which are currently not inferior to other tours in Indonesia. but don't forget the current covid 19 pandemic season, we have to be careful to keep our distance, use a mask, don't forget to always bring a hand sanitizer. The food that my family and I like when visiting the tour is meatballs. These are delicious food and when they are mixed together, they are very delicious. Besides that, the food in Lampung is very diverse, ranging from cheap to expensive, everything is there. I really like it when the holidays arrive my children and my wife love to go out of the house to calm hearts and minds.
Every day I work in one of the financial fields at a bank in my area. I work as a staff in the company that I work for until now. I spend eight hours a day working the job that I am currently doing is very enjoyable. I have worked for 8 years for a very long time from junior staff to senior staff. yes that's my life journey. We should always enjoy life, even though it's a little bit but it's fun, don't be sorry, just live this life full of enthusiasm and always struggle

I know about this community from my friend who is named @crismanto. I hope that in this hive community I can meet and exchange ideas with community friends. thanks